Flatten component structure
