+++ weather.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 from pytz import timezone, utc
 import logging
 
-from homeassistant.components import daikinskyport
+from . import DATA_DAIKINSKYPORT
 from homeassistant.components.weather import (
     ATTR_FORECAST_CONDITION,
     ATTR_FORECAST_TEMP,
@@ -28,10 +28,10 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
     if discovery_info is None:
         return
     dev = list()
-    data = daikinskyport.NETWORK
+    data = hass.data.get(DATA_DAIKINSKYPORT)
     for index in range(len(data.daikinskyport.thermostats)):
         thermostat = data.daikinskyport.get_thermostat(index)
-        dev.append(DaikinSkyportWeather(thermostat["name"], index))
+        dev.append(DaikinSkyportWeather(data, thermostat["name"], index))
 
     add_entities(dev, True)
 
@@ -39,8 +39,9 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
 class DaikinSkyportWeather(WeatherEntity):
     """Representation of Daikin Skyport weather data."""
 
-    def __init__(self, name, index):
+    def __init__(self, data, name, index):
         """Initialize the Daikin Skyport weather platform."""
+        self._data = data
         self._name = name
         self._index = index
         self.weather = None
@@ -108,7 +109,7 @@ class DaikinSkyportWeather(WeatherEntity):
 
     def update(self):
         """Get the latest state of the sensor."""
-        data = daikinskyport.NETWORK
+        data = self._data
         data.update()
         self.weather = dict()
         thermostat = data.daikinskyport.get_thermostat(self._index)
